See #6.  Add option to operate on integer dimensions.
We implement this somewhat simplistically: for integer dimensions, we round
the step to the nearest integer away from 0.  The stepsize for integers are
still halved (or doubled) with the step sizes for the rest of the dimensions.

Recall that the stencil that generates the steps for the trial points also
contains the number of halvings to perform if that point is accepted.  If the
step for a trial point is only along integer dimensions, then we no longer
perform any halvings if that point is accepted (although it may still double).

Integer values are stored as floats with the rest of the coordinates, which I
think is fine.

I don't know how robustly this will behave.  I guess we'll see.  To improve
robustness, the suggestion remains the same: increase the queue size.