Create a simple client to run the algorithm in serial.  Partially solves #12.
This gives us a simple API for the client object.  We wrap
`dask.distributed.Client` to conform to this API.

`client=` is now an optional parameter to search.  If not provided, then the
SerialClient will be used.

Also, the default `max_queue_size` was changed to 4 times the number of
dimensions, or num_threads + num_workers (whichever is greater).  This seems
like a reasonable default.